add `driver` feature
