+++ Cargo.toml
@@ -14,6 +14,10 @@ documentation = "https://docs.rs/async-io"
 keywords = ["mio", "epoll", "kqueue", "iocp", "wepoll"]
 categories = ["asynchronous", "network-programming", "os"]
 
+[features]
+default = ["driver"]
+driver = []
+
 [dependencies]
 concurrent-queue = "1.2.2"
 fastrand = "1.3.5"
